api/pyproject.toml
[tool.poetry]
name = "pear_template"
version = "0.1.0"
description = "A scaffold for Pear Code software development (and anywhere else)"
authors = ["Robert Townley"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.2"
psycopg2 = "^2.8.4"
easy-thumbnails = "^2.7"
djangorestframework = "^3.11.0"
django-anymail = {version = "^7.0.0", extras = ["mailgun"]}
django-cors-headers = "^3.2.1"
factory_boy = "^2.12.0"
django-rest-auth = "^0.9.5"
flake8 = "^3.7.9"
djoser = "^2.0.3"

[tool.poetry.dev-dependencies]
django-extensions = "^2.2.6"
django-webpack-loader = "^0.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
